cmd, node: begin migrating make node
