+++ cmd/gossamer/configcmd.go
@@ -25,25 +25,16 @@ import (
 	"strings"
 	"unicode"
 
-	"github.com/ChainSafe/gossamer/common"
 	cfg "github.com/ChainSafe/gossamer/config"
-	"github.com/ChainSafe/gossamer/config/genesis"
-	"github.com/ChainSafe/gossamer/core"
 	"github.com/ChainSafe/gossamer/internal/api"
-	"github.com/ChainSafe/gossamer/internal/services"
 	"github.com/ChainSafe/gossamer/keystore"
-	"github.com/ChainSafe/gossamer/network"
 	"github.com/ChainSafe/gossamer/node/gssmr"
-	"github.com/ChainSafe/gossamer/rpc"
-	"github.com/ChainSafe/gossamer/rpc/json2"
-	"github.com/ChainSafe/gossamer/runtime"
-	"github.com/ChainSafe/gossamer/state"
 	log "github.com/ChainSafe/log15"
 	"github.com/naoina/toml"
 	"github.com/urfave/cli"
 )
 
-// makeNode sets up node; opening badgerDB instance and returning the Node container
+// makeNode reads the configuration and makes the node
 func makeNode(ctx *cli.Context) (*gssmr.Node, *cfg.Config, error) {
 	currentConfig, err := getConfig(ctx)
 	if err != nil {
@@ -52,18 +43,7 @@ func makeNode(ctx *cli.Context) (*gssmr.Node, *cfg.Config, error) {
 
 	log.Info("🕸\t Configuring node...", "datadir", currentConfig.Global.DataDir, "protocol", currentConfig.Network.ProtocolID, "bootnodes", currentConfig.Network.Bootnodes)
 
-	var srvcs []services.Service
-
 	dataDir := currentConfig.Global.DataDir
-
-	// Create service, initialize stateDB and blockDB
-	stateSrv := state.NewService(dataDir)
-	srvcs = append(srvcs, stateSrv)
-
-	err = stateSrv.Start()
-	if err != nil {
-		return nil, nil, fmt.Errorf("cannot start db service: %s", err)
-	}
 
 	// load all static keys from keystore directory
 	ks := keystore.NewKeystore()
@@ -75,24 +55,6 @@ func makeNode(ctx *cli.Context) (*gssmr.Node, *cfg.Config, error) {
 		}
 	}
 
-	// Trie, runtime: load most recent state from DB, load runtime code from trie and create runtime executor
-	rt, err := loadStateAndRuntime(stateSrv.Storage, ks)
-	if err != nil {
-		return nil, nil, fmt.Errorf("error loading state and runtime: %s", err)
-	}
-
-	// load genesis from JSON file
-	gendata, err := stateSrv.Storage.LoadGenesisData()
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// TODO: Configure node based on Roles #601
-
-	// Network
-	networkSrvc, networkMsgSend, networkMsgRec := createNetworkService(currentConfig, gendata, stateSrv)
-	srvcs = append(srvcs, networkSrvc)
-
 	// BABE authority configuration; flag overwrites config option
 	if auth := ctx.GlobalBool(AuthorityFlag.Name); auth && !currentConfig.Global.Authority {
 		currentConfig.Global.Authority = true
@@ -102,46 +64,8 @@ func makeNode(ctx *cli.Context) (*gssmr.Node, *cfg.Config, error) {
 
 	log.Info("node", "authority", currentConfig.Global.Authority)
 
-	// Core
-	coreConfig := &core.Config{
-		BlockState:    stateSrv.Block,
-		StorageState:  stateSrv.Storage,
-		Keystore:      ks,
-		Runtime:       rt,
-		MsgRec:        networkMsgSend, // message channel from network service to core service
-		MsgSend:       networkMsgRec,  // message channel from core service to network service
-		BabeAuthority: currentConfig.Global.Authority,
-	}
-	coreSrvc := createCoreService(coreConfig)
-	srvcs = append(srvcs, coreSrvc)
-
-	// API
-	apiSrvc := api.NewAPIService(networkSrvc, nil)
-	srvcs = append(srvcs, apiSrvc)
-
-	// TODO: check rpc flag
-	rpcSrvr := startRPC(ctx, currentConfig.RPC, apiSrvc)
-
-	return gssmr.NewNode(gendata.Name, srvcs, rpcSrvr), currentConfig, nil
-}
-
-func loadStateAndRuntime(ss *state.StorageState, ks *keystore.Keystore) (*runtime.Runtime, error) {
-	latestState, err := ss.LoadHash()
-	if err != nil {
-		return nil, fmt.Errorf("cannot load latest state root hash: %s", err)
-	}
-
-	err = ss.LoadFromDB(latestState)
-	if err != nil {
-		return nil, fmt.Errorf("cannot load latest state: %s", err)
-	}
-
-	code, err := ss.GetStorage([]byte(":code"))
-	if err != nil {
-		return nil, fmt.Errorf("error retrieving :code from trie: %s", err)
-	}
-
-	return runtime.NewRuntime(code, ss, ks)
+	// TODO: make node based on implementation
+	return gssmr.MakeNode(ctx, currentConfig, ks)
 }
 
 // getConfig checks for config.toml if --config flag is specified and sets CLI flags
@@ -231,57 +155,6 @@ func setNetworkConfig(ctx *cli.Context, fig *cfg.NetworkCfg) {
 	}
 }
 
-// createNetworkService creates a network service from the command configuration and genesis data
-func createNetworkService(fig *cfg.Config, gendata *genesis.GenesisData, stateService *state.Service) (*network.Service, chan network.Message, chan network.Message) {
-	// Default bootnodes and protocol from genesis file
-	bootnodes := common.BytesToStringArray(gendata.Bootnodes)
-	protocolID := gendata.ProtocolID
-
-	// If bootnodes flag has one or more bootnodes, overwrite genesis bootnodes
-	if len(fig.Network.Bootnodes) > 0 {
-		bootnodes = fig.Network.Bootnodes
-	}
-
-	// If protocol id flag is not an empty string, overwrite
-	if fig.Network.ProtocolID != "" {
-		protocolID = fig.Network.ProtocolID
-	}
-
-	// network service configuation
-	networkConfig := network.Config{
-		BlockState:   stateService.Block,
-		StorageState: stateService.Storage,
-		NetworkState: stateService.Network,
-		DataDir:      fig.Global.DataDir,
-		Roles:        fig.Global.Roles,
-		Port:         fig.Network.Port,
-		Bootnodes:    bootnodes,
-		ProtocolID:   protocolID,
-		NoBootstrap:  fig.Network.NoBootstrap,
-		NoMdns:       fig.Network.NoMdns,
-	}
-
-	networkMsgRec := make(chan network.Message)
-	networkMsgSend := make(chan network.Message)
-
-	networkService, err := network.NewService(&networkConfig, networkMsgSend, networkMsgRec)
-	if err != nil {
-		log.Error("Failed to create new network service", "err", err)
-	}
-
-	return networkService, networkMsgSend, networkMsgRec
-}
-
-// createCoreService creates the core service from the provided core configuration
-func createCoreService(coreConfig *core.Config) *core.Service {
-	coreService, err := core.NewService(coreConfig)
-	if err != nil {
-		log.Error("Failed to create new core service", "err", err)
-	}
-
-	return coreService
-}
-
 func setRPCConfig(ctx *cli.Context, fig *cfg.RPCCfg) {
 	// Modules
 	if mods := ctx.GlobalString(RPCModuleFlag.Name); mods != "" {
@@ -298,13 +171,6 @@ func setRPCConfig(ctx *cli.Context, fig *cfg.RPCCfg) {
 		fig.Port = uint32(port)
 	}
 
-}
-
-func startRPC(ctx *cli.Context, fig cfg.RPCCfg, apiSrvc *api.Service) *rpc.HTTPServer {
-	if ctx.GlobalBool(RPCEnabledFlag.Name) {
-		return rpc.NewHTTPServer(apiSrvc.API, &json2.Codec{}, fig.Host, fig.Port, fig.Modules)
-	}
-	return nil
 }
 
 // strToMods casts a []strings to []api.Module
